deleted parsefloats

+++ src/models/Distillation/Distillation.ts
@@ -82,9 +82,9 @@ export class Distillation {
         const resultMap: {[key: string]: number} = {}
         Object.keys(data).forEach(year => {
             if(propToSum === SumByTypes.HLF){
-                resultMap[year] = data[year].reduce((acc: number, curr: Distillation) => acc + parseInt(curr.HLF), 0)
+                resultMap[year] = data[year].reduce((acc: number, curr: Distillation) => acc + parseFloat(curr.HLF), 0)
             } else {
-                resultMap[year] = data[year].reduce((acc: number, curr: Distillation) => acc + parseInt(curr.weightInKilograms), 0)
+                resultMap[year] = data[year].reduce((acc: number, curr: Distillation) => acc + parseFloat(curr.weightInKilograms), 0)
             }
         })
         return resultMap;
